Fix "Non ASCII character" SyntaxError exception
The exception is thrown by the non ASCII character "❄" used in the
"cli.py" (1) log message. Regarding to PEP-263 (2) the default encoding
is set to ASCII, but the file must be parsed as UTF-8.

This has been fixed by adding the "coding" magic commment to ensure UTF-8 encoding.

References:
  (1) https://github.com/arcticicestudio/snowsaw/blob/develop/snowsaw/cli.py#L87
  (2) https://www.python.org/dev/peps/pep-0263

GH-20