Add per-host capability to shell plugin
Currently, any configured script will execute on all hosts. This commit
will allow for a host item in the config so that the script will only
run on the specific host. A hostname of "-" will match all hosts. For
backwards compatibility, a config with no host specified will run
on all hosts.