Implement Prettier configuration and ignore pattern file
This is one of the main features of Prettier: It already provides the
best and recommended style configurations of-out-the-box.
The only option that has been changed is the print width (1). It is set
to `80` by default which not up-to-date for modern screens (might only
be relevant when working in terminals only like e.g. with Vim). It is
now set to `120` used by all of Arctic Ice Studio's style guides.
The `prettier.config.js` configuration file is placed in the project
root as well as the `.prettierignore` file that defines ignore pattern.

References:
  (1) prettier.io/docs/en/options.html#print-width

Epic: GH-33
Depends on GH-47 GH-49
GH-37