Renewed project documentations and assets
The initial content of the renewed project repository `README.md` and
`CHANGELOG.md` files now contain the SVG hero and badges provided by the
great shields.io (1) project. They make use to the available query
parameters to style and customize the badges which come closest to
Arctic Ice Studio's design. To render the SVG on GitHub the
`sanitize=true` query parameter is used to allow rendering SVG's hosted
within the repository and served via the `raw.githubusercontent`
subdomain.

They provide information about

- the latest GitHub release version and changelog
- Circle CI build and Codecov status
- used style guide versions

Additionally the README provides information how to contribute to the
project and the CHANGELOG includes a template for future version
entries.

Further documentations about the design concept, architecture and
technologies as well as guides for contributions to develop, run and
maintain the project will be documented later on.

References:
  (1) https://shields.io

Epic GH-33
Resolves GH-57