Go Module Initialization
snowsaw will be build using the latest Go version 1.12.6 (1) (to the
time this commit is written) with Go Modules (2).
This commits initialized snowsaw as Go module by running the
`go mod init` command in order to create the `go.mod` file.
The name of the module is `github.com/arcticicestudio/snowsaw` since
this repository will be used as origin hosted on GitHub.

References:
  (1) https://golang.org/dl/#go1.12.6
  (2) https://github.com/golang/go/wiki/Modules

Epic GH-33
Blocks GH-59 GH-60 GH-61 GH-62
Resolves GH-58