Basic level-based printer
To print any data/information, this commit implements a custom printer
that can be configured with a severity/level to control the verbosity
for different purposes.

The printer is located in the `pkg/prt` package that provides multiple
print formatting functions including some colored prefix symbols/words
for each level. This ensures the severity of the output can be
recognized at a glance.

Epic GH-33
Depends on GH-58
Resolves GH-59