App config handling with JSON/YAML de/encoder and builder
The `pkg/config` package provides the application-wide configuration
struct and constants.
The subpackage `pkg/config/encoder` provides the `Encoder` interface
that is implemented in the  `pkg/config/encoder/json` and
`pkg/config/encoder/yaml` packages to handle the de/encoding of JSON and
YAML data. To process the data from files the `File` struct of the
`pkg/config/source/file` package represents an abstractation to validate
and load data from files on the local filesystem.

To simplify the usage and handling of configurations the
`pkg/config/builder` package provides the `Load` and `Merge` functions
by using the builder design pattern in order to allow to chain both
functions.

References:
  (1)
  (2)

Epic GH-33
Depends on GH-58
Resolves GH-60