CLI interface setup and main/root command
The root command `snowsaw` represents the main entry point of the
application that

- registers all subcommands and global flags.
- initializes the application logging/printer (GH-59) verbosity level.
- checks for existing application-wide configuration files to load and
  merges them with the given flag parameters.

The `info` subcommand prints more detailed application information
while the `--version` flag can be used to obtain the version number in a
parsable format.

The `main` function in the `main` package has been placed in the
`main.go` in the repository root. It calls `Run()` of the `snowsaw`
command to start the main application flow.

Epic GH-33
Depends on GH-58
Resolves GH-61