Exported utility function packages
There was only the `internal/util` package, but the functions were well
suited for task runners that can be implemented by users to extend
snowsaw's capability to handle custom tasks.
To allow to use the utility functions the package has been moved to the
`pkg/util` parent package.

Epic GH-33
Resolves GH-66